Correcting warnings in websocketpp sources

+++ contrib/websocketpp/websocketpp/base64/base64.hpp
@@ -52,8 +52,7 @@ static inline bool is_base64(unsigned char c) {
            (c >= 97 && c <= 122)); // a-z
 }
 
-inline std::string base64_encode(unsigned char const * bytes_to_encode, unsigned
-    int in_len)
+inline std::string base64_encode(unsigned char const * bytes_to_encode, size_t in_len)
 {
     std::string ret;
     int i = 0;
@@ -103,7 +102,7 @@ inline std::string base64_encode(unsigned char const * bytes_to_encode, unsigned
 }
 
 inline std::string base64_encode(std::string const & data) {
-    return base64_encode(reinterpret_cast<const unsigned char *>(data.data()),data.size());
+    return base64_encode(reinterpret_cast<const unsigned char *>(data.data()), data.size());
 }
 
 inline std::string base64_decode(std::string const & encoded_string) {
